[FEAT/#259] exploreSearch 폴더 위치 변경
